split up Dokkatoo plugin into an un-opinionated plugin, and one that applies all format plugins
