+++ modules/dokkatoo-base-plugin/src/main/kotlin/tasks/DokkatooTask.kt
@@ -1,6 +1,6 @@
 package dev.adamko.dokkatoo.tasks
 
-import dev.adamko.dokkatoo.DokkatooPlugin
+import dev.adamko.dokkatoo.DokkatooBasePlugin
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.CacheableTask
 
@@ -9,7 +9,7 @@ import org.gradle.api.tasks.CacheableTask
 abstract class DokkatooTask : DefaultTask() {
 
   init {
-    group = DokkatooPlugin.TASK_GROUP
+    group = DokkatooBasePlugin.TASK_GROUP
   }
 
 }
